skb: Don't clear metadata in pskb_expand_head()
pskb_expand_head() expands the head buffer of an skb. Currently it
clears any metadata stored in the headroom, by clearing
skb_shared_info->meta_len.

To allow metadata to be accessed higher up the stack, we want to
preserve it.

There's no need to clear the metadata in pskb_expand_head(): the
headroom and data (just not the tailroom) are copied into the new head.