trait: limited KV store for packet metadata
A (very limited) KV store to support storing KVs in the packet headroom,
with:
- 64 keys (0-63).
- 2, 4 or 8 byte values.
- O(1) lookup
- O(n) insertion
- A fixed 16 byte header.

Values are stored ordered by key, immediately following the fixed
header.

This could be extended in the future, for now it implements the smallest
possible API.

For example we could support more value sizes:
- A "0" size, to indicate a key has been set with no value.
- 16 bytes to store IPv6 addresses.

And could also support more than 64 keys.

Implemented in the header file so functions are always inlinable.

Signed-off-by: Arthur Fabre <afabre@cloudflare.com>