trait: XDP support
Store the trait KV store in the xdp_buff headroom, right after the
xdp_frame.

Support for traits is gated on support for metadata: to propagate the
traits to the skb layer, drivers will have to explicitly signal support
via the existing skb_metadata_set() call.

This also makes bounds checks simpler - we can rely on xdp->data_meta
always being a valid hard_end for traits.

Error handling is not very graceful in the case there's not enough
headroom to store the traits header (16 bytes), but that should be
impossible in current drivers with an XDP program attached.