trait: Replace memcpy calls with inline copies
When copying trait values to or from the caller, the size isn't a
constant so memcpy() ends up being a function call.

Replace it with an inline implementation that only handles the sizes we
support.

We store values "packed", so they won't necessarily be 4 or 8 byte
aligned.

TODO - benchmark to see if this makes it faster.

Signed-off-by: Arthur Fabre <afabre@cloudflare.com>