trait: Support sk_buffs
Hide the space used by traits from skb_headroom(): that space isn't
actually usable.

Preserve the trait store in pskb_expand_head() by copying it ahead of
the new headroom. The struct xdp_frame at the start of the headroom
isn't needed anymore, so we can overwrite it with traits, and introduce
a new flag to indicate traits are stored at the start of the headroom.

Don't allow cloned skbs to access traits: changes would be reflected in
the original skb (as they share the same packet data and headroom)
which is probably not what a user expects.
Longer term we could support this by allowing read only access, and on
writes copying the traits to a skb extension.

Signed-off-by: Arthur Fabre <afabre@cloudflare.com>